Fix: 채팅 방에 접속 중이어도 안 읽은 채팅 개수 증가하는 점 수정
